Add fallible version of composer registration
This commit changes CheckComposerMap and CheckComposerWrapper so that
they natively handle the possibility of composable checks failing to
register.  An example, as given in the TicTacToe adapter, is when a
composable check pulls in an assertion that contradicts a previous
registered check.

When composers in wrappers or maps fail, they are effectively locked
into a failure state - further attempts to register checks with them will
be ignored.  This means that composer logic doesn't have to worry about
handling error handling at any point other than registration.

This should mostly be a backwards-compatible change: the fallible forms
of `register` are called `tryRegister`.  The `inner` property on
`CheckComposerWrapper` has changed type to `Result<S>`, which is
technically breaking.